use 'uv run pytest' for CI
Instead of our $extras var be a list, we will write it as the combination
of --extra ? and no extra flags for uv. Then we can skip the two separate
install steps, and run uv using the --with and $extras flags.

uv build/publish does not have the same kind of check as `twine check` so
that has not been updated yet.